Updated LICENSE in source code files.
Created file for PlayerStatistics class and created Player alias for that class
